Changed all int data type of all properties from estimatedExpense to decimal

+++ TravelApplication/Models/EstimatedExpense.cs
@@ -9,27 +9,27 @@ namespace TravelApplication.Models
     {
         public int EstimatedExpenseId { get; set; }
         public int TravelRequestId { get; set; }
-        public int AdvanceLodging { get; set; }
-        public int AdvanceAirFare { get; set; }
-        public int AdvanceRegistration { get; set; }
-        public int AdvanceMeals { get; set; }
-        public int AdvanceCarRental { get; set; }
-        public int AdvanceMiscellaneous { get; set; }
-        public int AdvanceTotal { get; set; }
-        public int TotalEstimatedLodge { get; set; }
-        public int TotalEstimatedAirFare { get; set; }
-        public int TotalEstimatedRegistration { get; set; }
-        public int TotalEstimatedMeals { get; set; }
-        public int TotalEstimatedCarRental { get; set; }
-        public int TotalEstimatedMiscellaneous { get; set; }
-        public int TotalEstimatedTotal { get; set; }
+        public decimal AdvanceLodging { get; set; }
+        public decimal AdvanceAirFare { get; set; }
+        public decimal AdvanceRegistration { get; set; }
+        public decimal AdvanceMeals { get; set; }
+        public decimal AdvanceCarRental { get; set; }
+        public decimal AdvanceMiscellaneous { get; set; }
+        public decimal AdvanceTotal { get; set; }
+        public decimal TotalEstimatedLodge { get; set; }
+        public decimal TotalEstimatedAirFare { get; set; }
+        public decimal TotalEstimatedRegistration { get; set; }
+        public decimal TotalEstimatedMeals { get; set; }
+        public decimal TotalEstimatedCarRental { get; set; }
+        public decimal TotalEstimatedMiscellaneous { get; set; }
+        public decimal TotalEstimatedTotal { get; set; }
         public string HotelNameAndAddress { get; set; }
         public string Schedule { get; set; }
         public string PayableToAndAddress { get; set; }
         public string Note { get; set; }
         public string AgencyNameAndReservation { get; set; }
         public string Shuttle { get; set; }
-        public int CashAdvance { get; set; }
+        public decimal CashAdvance { get; set; }
         public DateTime DateNeededBy { get; set; }
 
     }
